Refactor extract database query
